requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
bidict==0.23.1
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
diff-match-patch==20241021
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-formtools==2.5.1
django-import-export==4.3.9
django-leaflet==0.32.0
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
phonenumbers==8.13.55
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pypng==0.20220715.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
qrcode==7.4.2
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.0
setuptools==78.1.1
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.2
tablib==3.8.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
